requirements.txt
Flask==3.0.3
Flask-Uploads==0.2.1
torch==2.3.1
transformers==4.41.2
torchaudio==2.3.1
python-dotenv==1.0.1